fix of testMock
