+++ utils/testUtils.js
@@ -1,3 +1,5 @@
+const {generateMockRecipeData} = require("./testMock.js");
+
 /**
  * Tests for non-numeric user IDs by performing API requests and checking responses.
  *
@@ -215,54 +217,6 @@ function handleMultipleParamsTest(context) {
     );
   }
 }
-
-/**
- * Generates mock data for a recipe.
- * This function is primarily used for testing purposes, providing a consistent and static set of recipe data.
- * It includes details such as ingredients, nutritional information, instructions, and other relevant attributes.
- */
-const generateMockRecipeData = function (invalidParameter = false) {
-  let recipeData = {
-    IngredientsName: [
-      "Chicken Breast",
-      "Olive Oil",
-      "Garlic",
-      "Lemon Juice",
-      "Paprika",
-    ],
-    difficulty: "Easy",
-    servings: "5", // Intentionally incorrect type for testing
-    recipeType: "Main Course",
-    Instructions: [
-      "Preheat oven to 375°F (190°C).",
-      "In a bowl, mix olive oil, minced garlic, lemon juice, paprika, salt, and black pepper.",
-      "Place chicken breasts in a baking dish and pour the mixture over them.",
-    ],
-    NutritionalTable: [
-      "Calories: 165",
-      "Protein: 26g",
-      "Fat: 4.5g",
-      "Carbohydrates: 3g",
-      "Sodium: 322mg",
-    ],
-    IngredientsQuantity: [
-      "4 medium-sized chicken breasts",
-      "2 tablespoons olive oil",
-      "3 cloves garlic, minced",
-    ],
-    time: 5, // Assuming this should be a number
-    recipeID: 101,
-    imgRecipe: "https://example.com/images/chicken-breast-recipe.jpg",
-  };
-  // If invalidParameter is true, modify the servings field to an invalid value
-  if (!invalidParameter) {
-    recipeData.servings = "invalid_number`";
-  } else {
-    recipeData.servings = 5;
-  }
-  return recipeData;
-};
-
 
 // Helper function to generate a random test ID
 function generateTestID() {
